Verification: migrate to Epoxy - rename package
